# Monitor dogstatsd buffer usage