@olivielpeau 2020-09-24
+1 on benchmarking this. @remeh any thoughts on the easiest way to do so? (maybe a small benchmark test on `PacketsBuffer` only could be enough?)

Note that there can be more than one `PacketsBuffer` instantiated (for example: one for UDS and one for UDP), so the new telemetry metrics would be updated by both. Since they share the output channel, the new telemetry metrics would still make sense I think, but worth noting.