@remeh 2020-09-24
> +1 on benchmarking this. @remeh any thoughts on the easiest way to do so? (maybe a small benchmark test on `PacketsBuffer` only could be enough?)
> 
> Note that there can be more than one `PacketsBuffer` instantiated (for example: one for UDS and one for UDP), so the new telemetry metrics would be updated by both. Since they share the output channel, the new telemetry metrics would still make sense I think, but worth noting.

Two valid points!

 * a Go benchmark should be enough to validate the impact and I was thinking of maybe (after the release) deploying a custom build on a staging k8s cluster to observe the result as well. It is a good "small change" candidate to test this.
  * about the writing in the channel: you're right and as you mentioned, because it is shared the value should still make sense. However, that may be useful to add a new metric with the amount of created listeners? WDYT?