@olivielpeau 2020-09-24
> a Go benchmark should be enough to validate the impact and I was thinking of maybe (after the release) deploying a custom build on a staging k8s cluster to observe the result as well. It is a good "small change" candidate to test this.

ok! Let's add a benchmark on the `PacketsBuffer` to this PR to start with then, @PierreGuilmin could you look into it?

> However, that may be useful to add a new metric with the amount of created listeners? WDYT?

Not sure if it's worth sending as a metric, since it's a static value and on a given environment it's easy to tell if UDP and/or UDS are enabled 🤔 
Ideally, each metric here would be tagged by `listener` (`udp` or `uds`) so that we can tell which listener is the cause, but I'm not sure if tagging with `telemetry` could have a performance impact here.