@PierreGuilmin 2020-09-24
> ok! Let's add a benchmark on the PacketsBuffer to this PR to start with then, @PierreGuilmin could you look into it?

On it :wink:

> Ideally, each metric here would be tagged by listener (udp or uds) so that we can tell which listener is the cause, but I'm not sure if tagging with telemetry could have a performance impact in a hot path here.

I was concerned by performance regression as well, that's why I didn't add it. We can maybe evaluate the performance impact of adding this tag using the new benchmark and decide wether we should do it or not?