@PierreGuilmin 2020-09-25
@remeh @olivielpeau I just pushed the `PacketsBuffer` benchmark and updated the PR description to detail the  performance impact.
> So this PR causes a statistically significant (p < 0.05) increase of ~ **5%** computation time. 🚀

I think it is definitely not worth it to add tagging here.