@remeh 2020-09-25
> Tagging results in a ~ 20% computation time increase, but maybe the increase is very small over the whole pipeline (this is just the increase over three functions), it would be nice to have a global benchmark to see if adding tagging has a real impact.

That's a very interesting benchmark, thanks for having added it Pierre.