@remeh 2020-10-05
> 👋  PierreGuilmin has now finished his internship. What's left to do for this PR to be merged ?

@iksaif 
As it's in a hot path, we want to validate first that the added telemetry isn't introducing too much cpu usage., but except from that, this is ready to go. We've a card to track this work (AC-631).